# backend/openapi.yml
openapi: 3.0.3
info:
  title: WikiLive API
  version: 1.0.0
servers:
  - url: http://localhost:8085
    description: Local
  - url: https://wiki-live.ru
    description: Production
  - url: https://www.wiki-live.ru
    description: Production
security:
  - basicAuth: []
paths:
...
  /api/pages/{slug}/comments:
    get:
      tags:
        - Comments
      summary: Get comments for page
      security: []
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
            maxLength: 255
            pattern: "^[a-z0-9-]+$"
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CommentDTO'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        'default':
          $ref: '#/components/responses/Error'

    post:
      tags:
        - Comments
      summary: Add comment
      security: []
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
            maxLength: 255
            pattern: "^[a-z0-9-]+$"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentDTO'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        'default':
          $ref: '#/components/responses/Error'
  /api/pages/{slug}/comments/{id}:
    delete:
      tags:
        - Comments
      summary: Delete comment
      security: []
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
            maxLength: 255
            pattern: "^[a-z0-9-]+$"
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: No Content
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        'default':
          $ref: '#/components/responses/Error'
components:
  schemas:
...
    CommentDTO:
      type: object
      additionalProperties: false
      properties:
        id:
          type: integer
          format: int64
        author:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
    CommentRequest:
      type: object
      required:
        - author
        - content
      additionalProperties: false
      properties:
        author:
          type: string
        content:
          type: string
    ErrorResponse:
      type: object
      additionalProperties: false
      properties:
        message:
          type: string
        code:
          type: string
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Error:
      description: Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic